Limit material list to only visible objects
Checking Object.hide_get isn't enough to determine if an Object is hidden as it could also be disabled globally in the viewport (can be checked with Object.hide_viewport) or all the Collections the Object is in could be hidden either temporarily or globally, which would also hide the Object. Object.visible_get can be used to test if an Object is visible, taking into account all visibility settings, alternatively, given a screen Context, Context.visible_objects provides access to all Objects visible in that Context.
Only set Objects in the material list to OBJECT mode
Use context overrides to set Objects to OBJECT mode without changing the active Object
Fixes #45